+++ marshmallow/fields.py
@@ -474,10 +474,10 @@ class Number(Field):
         return self.num_type(value)
 
     def _validated(self, value):
-        """Format the value or raise ``exception_class`` if an error occurs."""
+        """Format the value or raise a :exc:`ValidationError` if an error occurs."""
         try:
             return self._format_num(value)
-        except (TypeError, ValueError, decimal.InvalidOperation) as err:
+        except (TypeError, ValueError) as err:
             raise ValidationError(getattr(self, 'error', None) or text_type(err))
 
     def serialize(self, attr, obj, accessor=None):
@@ -523,6 +523,8 @@ class Decimal(Number):
     :param rounding: How to round the value during quantize, for example
         `decimal.ROUND_UP`. If None, uses the rounding value from
         the current thread's context.
+    :param allow_nan: If `True`, `NaN`, `Infinity` and `-Infinity` are allowed, even
+        though they are illegal according to the JSON specification.
     :param default: The value this field defaults to. If not specified is the
         `decimal.Decimal` zero.
     :param bool as_string: If True, serialize to a string instead of a Python
@@ -534,18 +536,41 @@ class Decimal(Number):
 
     num_type = decimal.Decimal
 
-    def __init__(self, places=None, rounding=None, as_string=False, **kwargs):
+    def __init__(self, places=None, rounding=None, allow_nan=False, as_string=False, **kwargs):
         self.places = decimal.Decimal((0, (1,), -places)) if places is not None else None
         self.rounding = rounding
+        self.allow_nan = allow_nan
         super(Decimal, self).__init__(as_string=as_string, **kwargs)
 
+    # override Number
     def _format_num(self, value):
         if value is None:
             return None
+
         num = decimal.Decimal(value)
-        if self.places is not None:
+
+        if self.allow_nan:
+            if num.is_nan():
+                return decimal.Decimal('NaN')  # avoid sNaN, -sNaN and -NaN
+        else:
+            if num.is_nan() or num.is_infinite():
+                raise ValidationError(
+                    getattr(self, 'error', None) or 'Special numeric values are not permitted.'
+                )
+
+        if self.places is not None and num.is_finite():
             num = num.quantize(self.places, rounding=self.rounding)
+
         return num
+
+    # override Number
+    def _validated(self, value):
+        try:
+            return super(Decimal, self)._validated(value)
+        except decimal.InvalidOperation:
+            raise ValidationError(
+                getattr(self, 'error', None) or 'Invalid decimal value.'
+            )
 
 
 class Boolean(Field):
@@ -569,17 +594,19 @@ class Boolean(Field):
         try:
             value_str = text_type(value)
         except TypeError as error:
-            raise ValidationError(text_type(error))
+            msg = getattr(self, 'error', None) or text_type(error)
+            raise ValidationError(msg)
         if value_str in self.falsy:
             return False
         elif self.truthy:
             if value_str in self.truthy:
                 return True
             else:
-                raise ValidationError(
-                    '{0!r} is not in {1} nor {2}'.format(
-                        value_str, self.truthy, self.falsy
-                    ))
+                default_message = '{0!r} is not in {1} nor {2}'.format(
+                    value_str, self.truthy, self.falsy
+                )
+                msg = getattr(self, 'error', None) or default_message
+                raise ValidationError(msg)
         return True
 
 class FormattedString(Field):
@@ -694,16 +721,17 @@ class DateTime(Field):
         self.dateformat = format
 
     def _serialize(self, value, attr, obj):
-        if value:
-            self.dateformat = self.dateformat or self.DEFAULT_FORMAT
-            format_func = self.DATEFORMAT_SERIALIZATION_FUNCS.get(self.dateformat, None)
-            if format_func:
-                try:
-                    return format_func(value, localtime=self.localtime)
-                except (AttributeError, ValueError) as err:
-                    raise ValidationError(getattr(self, 'error', None) or text_type(err))
-            else:
-                return value.strftime(self.dateformat)
+        if value is None:
+            return None
+        self.dateformat = self.dateformat or self.DEFAULT_FORMAT
+        format_func = self.DATEFORMAT_SERIALIZATION_FUNCS.get(self.dateformat, None)
+        if format_func:
+            try:
+                return format_func(value, localtime=self.localtime)
+            except (AttributeError, ValueError) as err:
+                raise ValidationError(getattr(self, 'error', None) or text_type(err))
+        else:
+            return value.strftime(self.dateformat)
 
     def _deserialize(self, value):
         msg = 'Could not deserialize {0!r} to a datetime object.'.format(value)
@@ -745,6 +773,8 @@ class Time(Field):
     """
 
     def _serialize(self, value, attr, obj):
+        if value is None:
+            return None
         try:
             ret = value.isoformat()
         except AttributeError:
@@ -772,6 +802,8 @@ class Date(Field):
     """
 
     def _serialize(self, value, attr, obj):
+        if value is None:
+            return None
         try:
             return value.isoformat()
         except AttributeError:
@@ -824,6 +856,8 @@ class TimeDelta(Field):
         super(TimeDelta, self).__init__(error=error, **kwargs)
 
     def _serialize(self, value, attr, obj):
+        if value is None:
+            return None
         try:
             days = value.days
             if self.precision == self.DAYS: